migrate easy confirmatino to typescript
